Use git apply for updating projects and improve diff viewer
Instead of using `patch` cruft now uses `git apply -3` to update the
project. This means we can now handle renames and deletes. We also
get the same merge behavior in macOS and Linux for conflicting patches.
However this only works if the user's project directory is a git repo.
If not we fall back to git apply with a reject flag which generates
"*.rej" files which is similar to `patch` with `--no-backup-if-mismatch`
flag.

We also pip the diff to the git diff utility, which means users can
now view the patches with a pager and also use their default git diff
viewer to view the patches. For eg. they can use `diff-so-fancy`.

Fixes #21
Fixes #15
Fixes #33