Move CLI to Typer and improve test coverage to 100%
This PR moves the CLI to typer with which we now have proper CLI help
and also a slight update in terms of CLI options.
It also improves test coverage to a 100% once all the actions in all the
OS run (codecov allows you to combine coverage reports from different
environments! That is neat!). Apart from that this PR also makes user
dependency on examples optional.

Breaking changes with 1.x to note -

- The cruft api has now moved from `cruft.api` to `cruft`.
- Calling cruft via the API defaults to non-interactive flags of the
command where are with the CLI version, interactive flags are used.
- `use_latest` in `link` has been removed and replaced with `checkout`
which accepts branches, tags and commits and `link` always uses the
latest commit of the provided reference. This is much more useful
than users providing SHA hashes, they can now link against release
tags for eg while still maintain compatibility to link against a
particular commit.
- `cookiecutter_input` flag in `update` now only asks from inputs for
the new version of the template and uses `.cruft.json` context for the
currently linked version. This is to ensure that users don't modify the
context with which the template was linked during an `update`
- Not a breaking change but a change in behaviour - if directory is not
provided by user, it is now serialized as a null instead of a blank
string in cruft.json