Make cruft update more robust
This commit makes cruft update more robust when applying updates to a
project.

1. We terminate early if the project we are updating is a git repo
and the repo is unclean. This is to avoid git 3 way merge from failing
and the user applying changes via cruft update that they cannot recover.
This doesn't affect the user if the repo they are working with is not
a git repo.

2. When applying updates, if the user has deleted any files generated
from the cookiecutter template, we now ignore any updates to those
files. This makes the update more robust as git apply won't fail
and if the user has explicitly deleted any generated files, we shouldn't
try and update those.