Fix cruft update and change the diff order
This commit provides the following fixes -

https://github.com/cruft/cruft/issues/67
Fixes cruft not detecting new files

https://github.com/cruft/cruft/issues/53
Fixes cruft not being able to handle moves

Fixes cruft diff order so that the base repo is the project repo.
This way any changes from the upstream template can be outputted to
a diff file and applied if needed.