Fix diff crash on cookiecutter.json with _ prefix
Per #166, using "cruft diff" or "cruft update" on some templates causes
crash with obscure cookiecutter error like

    ValueError: ['jinja2_git.GitExtension'] provided for choice variable _extensions, but the choices are ['jinja2_git.GitExtension'].

or
    ValueError: ['.chglog/CHANGELOG.tpl.md', '.github/workflows/terraform.yml'] provided for choice variable _copy_without_render, but the choices are ['.chglog/CHANGELOG.tpl.md', '.github/workflows/terraform.yml'].

It seems that the cookiecutter.json file of those templates indisposes
cruft, as it tries to parse these as actual user inputs to replay.
But these are actually template-side extensions not meant for users to
set or modify, most of the time, and are defined by the JSON key
starting with an underscore: "_".

Fix the issue by preventing cruft from passing these variables (stored
in .cruft.json legitimately) as extra context = as pretend-user-input.
Default values for each of these keys will used instead as defined in
the cookiecutter.json file of the template.

Fixes #166