Accept json file for extra_context
Added get_extra_context_from_file in cookiecutter.py to parse the extra_context json and return it to the create command.

Added an extra_context_file option to _cli.py

If the create command is called it will first see if the extra_context_file exists and if so it will parse the file and assign it to the extra_context variable.